Restructure backend for testing
