+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "start": "node server",
-    "dev": "nodemon server"
+    "start": "node src/index",
+    "dev": "nodemon src/index"
   },
   "author": "",
   "license": "ISC",
