+++ src/api/v1/routes/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router()
+
+router.use('/auth', require('./auth'))
+router.use('/nfts', require('./nfts'))
+
+module.exports = router
