+++ src/api/v1/routes/nfts.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { getAllNFTs } = require('../controllers/nftController')
+const { getAllNFTs } = require('../controllers/nfts')
 
 router.get('/all', getAllNFTs)
